// client-master/src/app/models/Post.ts
export class Post {
  id: string;
  categoryId: string;
  accountId: string;
  province: string;
  district: string;
  ward: string;
  street: string;
  apartmentNumber: string;
  lat: number;
  lng: number;
  describe: string;
  utilities: string;
  images: string;
  acreage: number;
  direction: string;
  numberOfRooms: number;
  numberOfToliets: number;
  tolietType: string;
  status: string;
  rentalObject: string;
  closingTime: number;
  price: number;
  dateSubmitted: string;
  dateExpiration: string;
  isActive: boolean
}